Move ALPN dialer, ALPN conn upgrade, Ping conn to api
