+++ lib/web/conn_upgrade.go
@@ -25,17 +25,17 @@ import (
 	"github.com/gravitational/trace"
 	"github.com/julienschmidt/httprouter"
 
-	"github.com/gravitational/teleport"
+	"github.com/gravitational/teleport/api/constants"
 	"github.com/gravitational/teleport/lib/utils"
 )
 
 // selectConnectionUpgrade selects the requested upgrade type and returns the
 // corresponding handler.
 func (h *Handler) selectConnectionUpgrade(r *http.Request) (string, ConnectionHandler, error) {
-	upgrades := r.Header.Values(teleport.WebAPIConnUpgradeHeader)
+	upgrades := r.Header.Values(constants.WebAPIConnUpgradeHeader)
 	for _, upgradeType := range upgrades {
 		switch upgradeType {
-		case teleport.WebAPIConnUpgradeTypeALPN:
+		case constants.WebAPIConnUpgradeTypeALPN:
 			return upgradeType, h.upgradeALPN, nil
 		}
 	}
@@ -90,7 +90,7 @@ func (h *Handler) upgradeALPN(ctx context.Context, conn net.Conn) error {
 
 func writeUpgradeResponse(w io.Writer, upgradeType string) error {
 	header := make(http.Header)
-	header.Add(teleport.WebAPIConnUpgradeHeader, upgradeType)
+	header.Add(constants.WebAPIConnUpgradeHeader, upgradeType)
 	response := &http.Response{
 		Status:     http.StatusText(http.StatusSwitchingProtocols),
 		StatusCode: http.StatusSwitchingProtocols,
